+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.6.2
 - Improve back face selection culling edge / vert selection (fix bugs and more heuristics in ambiguous situations)
+- Fix regression: cancelling a transform modal causes UV warping
 
 # 1.6.1
 - Move gltf export items into the default blender gltf export; remove Anvil's separate scaled export